Made changes by moving my page components to the page folder and created a component for my NavBar and created a Navigation Layout
